CT-122 사용자 정보 수정 페이지 리팩토링
